+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/universal/UniversalSchemaPayloadDeserializer.scala
@@ -12,6 +12,7 @@ import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.confluent.serializ
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.serialization.GenericRecordSchemaIdSerializationSupport
 
 import java.nio.ByteBuffer
+import java.nio.charset.StandardCharsets
 
 trait UniversalSchemaPayloadDeserializer {
 
@@ -79,6 +80,20 @@ object JsonSchemaPayloadDeserializer extends UniversalSchemaPayloadDeserializer 
     val bytes = new Array[Byte](buffer.remaining())
     buffer.get(bytes)
     jsonSchema.deserializer.deserialize(bytes)
+  }
+
+}
+
+object PlainTextPayloadDeserializer extends UniversalSchemaPayloadDeserializer {
+
+  override def deserialize(
+      expectedSchemaData: Option[RuntimeSchemaData[ParsedSchema]],
+      writerSchemaData: RuntimeSchemaData[ParsedSchema],
+      buffer: ByteBuffer
+  ): Any = {
+    val bytes = new Array[Byte](buffer.remaining())
+    buffer.get(bytes)
+    new String(bytes, StandardCharsets.UTF_8)
   }
 
 }